+++ Task4/Task4/RetroGame/src/States/MainMenu/MainMenu.cpp
@@ -125,8 +125,8 @@ void MainMenuState::Draw()
 			int y = GetScreenHeight()/2 + (i*32);
 			if (j)
 			{
-				x += sinf(GetTime() * 8) * 4;
-				y += -cosf(GetTime() * 8) * 4;
+				x += (int)(sinf((float)GetTime() * 8) * 4);
+				y += (int)(-cosf((float)GetTime() * 8) * 4);
 			}
 			DrawText(optionText.str().c_str(), (int)m_menuOffset + x, y, 24, WHITE);
 		}
